requirements.txt
algorithmia>=1.0.0,<2.0

python-dotenv~=0.20.0